building matrix lib

+++ utils/functions.cpp
@@ -2,18 +2,18 @@
 
 
 // Sigmoid
-double sigmoid(double x) 
+float sigmoid(float x) 
 {
-    return 1.0 / (1.0 + exp(-x));
+    return 1.0f / (1.0f + exp(-x));
 }
 
-double sigmoid_derivative(double x)
+float sigmoid_derivative(float x)
 {
-    return sigmoid(x) * (1 - sigmoid(x));
+    return sigmoid(x) * (1.0f - sigmoid(x));
 }
 
 // ReLU
-double relu(double x)
+float relu(float x)
 {
     if (x > 0)
         {return x; }
@@ -21,7 +21,7 @@ double relu(double x)
         {return 0; }
 }
 
-double relu_derivative(double x)
+float relu_derivative(float x)
 {
     if (x > 0)
         {return 1; }
@@ -30,31 +30,17 @@ double relu_derivative(double x)
 }
 
 // Random Double Generator
-double random(double low, double high)
+float random(float low, float high)
 {
-    return low + static_cast<double> (rand()) / ( static_cast<double> (RAND_MAX / (high - low)) );
+    return low + static_cast<float>(rand()) / ( static_cast<float>(RAND_MAX / (high - low)) );
 }
 
 // Euclidean Distance
-double euclidean_distance(const std::vector<double>& vec1,  const std::vector<double>& vec2)
+float euclidean_distance(const std::vector<float>& vec1,  const std::vector<float>& vec2)
 {
-    double dist {0};
+    float dist {0};
     for (size_t idx {0}; idx < vec1.size(); idx++)
         { dist += pow(vec1[idx] - vec2[idx], 2); }
 
     return pow(dist, 0.5);
 }
-
-
-
-// int main()
-// {
-//     std::vector<double> v1 = {1., 2., 3.};
-//     std::vector<double> v2 = {4., 5., 6.};
-
-//     std::cout << euclidean_distance(v1, v2);
-    
-
-//     // std::cout << random(0.0, 1.0);
-// }
-
